refactor: Fix main class file name
