+++ packages/admin/src/firestore_simple.ts
