+++ packages/admin/src/index.ts
@@ -1,3 +1,3 @@
-export { FirestoreSimple } from './admin'
+export { FirestoreSimple } from './firestore_simple'
 export type { Collection } from './collection'
 export type { Encodable, Decodable } from './types'
