+++ packages/web/src/index.ts
@@ -1,4 +1,4 @@
 // Web SDK
-export { FirestoreSimple } from './web'
+export { FirestoreSimple } from './firestore_simple'
 export type { Collection } from './collection'
 export type { Encodable, Decodable } from './types'
